% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read fars data file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string or integer representing the location of
the file to be read.}
}
\value{
This function returns a tibble that includes the data in filename.
}
\description{
This is a function that returns a tibble based on a csv.bz2 file.
}
