% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize the data based on years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A list of integers representing the years to include.}
}
\value{
This function returns a tibble with the number of accidents grouped
   by year and month.
}
\description{
This is a function that reads multiple fars data files based on a list of 
years and provides the sum of data points for each year and month
}
\examples{
fars_summarize_years(2013)
fars_summarize_years(2013:2015)
fars_summarize_years(c(2013, 2015))

}
