% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create file}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A character string representing the year.}
}
\value{
This function returns the filename that corresponds to year.
}
\description{
This is a function that creates a filename based on the year.
}
